Fixed the kcalcore dependency. Modified the location of the temporary files creation to ~/.cache. Added a work around to re-initialize QContactManager at the time of which, without which msyncd just exits Fixed indentation. Removed the hack to create a new contacts manager
Made the necessary changes to be in sync with the directory and library
structure changes in buteo-syncfw
Fixed bug in the copying of /etc/buteo/sync files
Updated the .pro files to use pkgconfig of buteosyncfw rather than
hardcoding the library and include paths

Updated the xml files of the plug-ins to reflect the changes done
to the profile xmls in buteosyncfw. The service and the sync profile
xml files are merged into a single sync xml file.
To being with, change is done only to memotoo.xml. Once proper testing
for this is complete, change would be done to other profiles too
Modified the location of the buteo plugins directory to /usr/lib/buteo-plugins
Fixed issues with include paths

Signed-off-by: Sateesh Kavuri <sateesh@mobilitas.co>