Initial structure of the server plugin
To handle both bluetooth and USB connectors, USBConnection and BTConnection
classes would be created. The USB connector would require a USB kernel module
to create a /dev/ttyGS1 (possibly a gadget driver) that can handle the communication
with the PC host. The PC host can be an application that talks SyncML and will
act as a SyncML client, which this plugin will act as a OBEX server as
well as SyncML server.

Signed-off-by: Sateesh Kavuri <sateesh@mobilitas.co>