Created server plugin xml and modified the .pro file accordingly
Created the server plugin xml for USB connectivity and modified the .pro
file to copy the to the target dir. Also fixed issues with USBConnection
inherinting from QObject. Apparently, QObject has to be the first class
to inherit from, rather than OBEXConnection. Tested the server plugin
with syncml-ds-tool (from OpenSync project) over USB and could see that
sync invocation happens. But the syncml-ds-tool exits after sometime out
because of which the server plugin complains of link failure. This needs
to be debugged further. Also, there is an issue with QSocketNotifier
that the 'activated' signal gets triggered quite frequently. This
probably needs to be replaced with glib socket observer functionality

Signed-off-by: Sateesh Kavuri <sateesh@mobilitas.co>