Used glib monitoring of USB fd
Using QSocketNotifier is causing multiple notifications. Also fine
handling of the connects/disconnects does not seem to be possible (or
atlest, I need to know how to do it). Added a DEFINES flag to switch
between QSocketNotifier and glib. Whenever I am able to figure out
correct usage of QSocketNotifier (or whenever QSocketNotifier impl is
fixed), I will remove usage of glib. With this patch, sync works fine.
After every sync session, the fd gets closed (because of the way SyncML
stack is implemented). Is this ok?

Signed-off-by: Sateesh Kavuri <sateesh@mobilitas.co>