Added Qt way of listening for the fd
The Qt way of listening for fd is to use QSocketNotifier. But it does
not look as fine-grained as glib way. With this patch, sync works, but
there are warnings related to using QObject::connect and
QObject::disconnect methods from different threads. Need to handle these
Enable glib fd watching for now

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>