Fixed issue with connecting signals of QSocketNotifiers from different threads
QSocketNotifier belongs to thread ServerRunner while USBConnection
belongs to the transport thread in buteo syncml stack. So it is not
directly possible to connect signal/slots of these objects. The valid
way is to use the 5'th argument of the QObject::connect to specify
Qt::BlockingQueuedConnection instead of the default

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>