Added functionality to add SyncML client and server bluetooth records
Bluetooth profiles are made discoverable by bluetooth record called
"Service Discovery Record". As per Bluez way, the service responsible
for serving the specific service should add the record when the service
is live. This ensures that the external party is able to discover as
well as immediately connect to the service. In this case, the SyncML
server plugin adds these records and removes them at the time of
unloading of the service

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>