First cut implementation of BT connectivity for server plugin ready
A BT connection is created over an RFComm socket (which follows similar
semantics to opening a tcp/ip socket). This socket is then passed onto
the SyncML protocol via the ::connect() interface method. QSocketNotifer
is used to keep track of incoming,outgoing and errors over the BT
socket. Semantics similar to USB transport are used for BT transport as
well.

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>