Added the accept() call to handle the incoming BT data connection
The accept() call which would handle the incoming data connection of the
peer was missing. Added this call and set the mFd to be the fd of the
accept call. Only a BT SyncML server connection (channel=26) is good
enough in my view for a remote device to synchronize with the client. It
might be a different case for device-device sync, for which a client
channel might also be requird.

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>