Added support for device-device sync over bluetooth
This patch enables device-device synchronization. Earlier, the server
plugin was listening only on the BT server socket. Now a client BT
socket is also opened and the plugin listens on the socket for incoming
connections over client profile. Tested this with N9 and all the
contacts have been synchronized. This patch should also enable
synchronization between two Sailfish devices, but that needs to be
tested (when I have one more working sailfish device)

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>