Functionality to fetch sync settings from Accounts
The sync settings would be initially fetched from the sync profile. If
an account exists for the corresponding profile, then the settings would
be fetched from the account and they would be merged with the profile
settings. For this to work, the 'key' names in profile and account
settings should be the same. Account settings should use the keys in
ButeoProfileSettings.h header

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>