Added support for fetching credentials and settings from Accounts&SSO
Only support for "password" method and mechanism has been added to fetch
credentials. When there is further need for supporting SSO (whenever
such SyncML service is added), support can be added for that too. If
use_accounts=true in the profile, only then credentials and settings are
retrieved from SSO and Accounts. Else, the settings from profile would
be used. Settings the credentials in profile should be used only for
testing purposes

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>