Modified the syncmlclient.pro file to generate a OOP plugin
In order to support the generation of a binary exe instead of a dll, the
syncmlclient.pro file is modified to add a new config to generate an
exe. This exe config makes use of the plugin_main.cpp,
PluginServiceObj.h and PluginServiceObj.cpp which have the shell
functionality to communicate with msyncd. No other code changes would be
required to modify a dll plugin to an exe plugin

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>