[buteo-sync-plugins] Use remote_name instead of remote_id for OBEX
Some Bluetooth OBEX sync profiles don't have a remote_id value set
but only a remote_name.  This commit ensures that the remote_name
value is used as the target id when initialising the config for the
profile, allowing it to be removed (as the cleanUp() function in
the syncML library requires that value to be non-empty for remove).